// frontend/src/lib/api.d.ts
export interface components {
    schemas: {
// ...
        BucketCreate: {
            description?: string | null;
            /** @default 1 */
            history?: number;
            /** @default -1 */
            max_bytes?: number;
            /** @default -1 */
            max_value_size?: number;
            name: string;
            /** @default 1 */
            replicas?: number;
            storage?: components["schemas"]["Storage"];
            ttl_seconds?: number | null;
        };
// ...
        KvKeyRow: {
            /** Format: date-time */
            created: string;
            key: string;
            operation: components["schemas"]["KvOperation"];
            revision: number;
            size: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}